+++ packages/core/modal-dialog/src/components/__tests__/unit/server-side.js
@@ -5,13 +5,27 @@
 import React from 'react';
 import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
 import ReactDOMServer from 'react-dom/server';
+import Modal from '../../../../src';
 
-/* AK-5590: SSR tests failing because of window is not defined */
-/* eslint-disable jest/no-disabled-tests */
-test.skip('Modal dialog server side rendering', async () => {
+test('Modal dialog server side rendering', async () => {
   (await getExamplesFor('modal-dialog')).forEach(examples => {
     // $StringLitteral
     const Example = require(examples.filePath).default; // eslint-disable-line import/no-dynamic-require
     expect(() => ReactDOMServer.renderToString(<Example />)).not.toThrowError();
   });
+});
+
+test('Modal dialog should render content in ssr', () => {
+  const Content = jest.fn(() => 'Hello');
+  const modalString = ReactDOMServer.renderToString(
+    <Modal
+      heading="Look at this"
+      actions={[{ text: 'Close', onClick: () => {} }]}
+    >
+      <Content />
+    </Modal>,
+  );
+  expect(Content).toHaveBeenCalled();
+  expect(modalString).toContain('Close');
+  expect(modalString).toContain('Look at this');
 });