+++ packages/editor/editor-common/src/__tests__/unit/json-schema/v1.ts
@@ -1,24 +1,41 @@
 import { name } from '../../../../package.json';
-import * as v1schema from '../../../../json-schema/v1/full.json';
 import * as Ajv from 'ajv';
 import { readFilesSync } from '../../../../test-helpers';
 
+import * as v1schemaFull from '../../../../json-schema/v1/full.json';
+import * as v1schemaStage0 from '../../../../json-schema/v1/stage-0.json';
+
 const ajv = new Ajv();
-const validate = ajv.compile(v1schema);
+
+const schemas = {
+  full: v1schemaFull,
+  'stage-0': v1schemaStage0,
+};
 
 describe(`${name} json-schema v1`, () => {
-  const valid = readFilesSync(`${__dirname}/v1-reference/valid`);
-  valid.forEach(file => {
-    it(`validates '${file.name}`, () => {
-      validate(file.data);
-      expect(validate.errors).toEqual(null);
-    });
-  });
+  Object.keys(schemas).forEach(schemaName => {
+    describe(schemaName, async () => {
+      const schema = schemas[schemaName];
+      const validate = ajv.compile(schema);
 
-  const invalid = readFilesSync(`${__dirname}/v1-reference/invalid`);
-  invalid.forEach(file => {
-    it(`does not validate '${file.name}`, () => {
-      expect(validate(file.data)).toEqual(false);
+      const valid = readFilesSync(
+        `${__dirname}/v1-reference/${schemaName}/valid`,
+      );
+      valid.forEach(file => {
+        it(`validates '${file.name}`, () => {
+          validate(file.data);
+          expect(validate.errors).toEqual(null);
+        });
+      });
+
+      const invalid = readFilesSync(
+        `${__dirname}/v1-reference/${schemaName}/invalid`,
+      );
+      invalid.forEach(file => {
+        it(`does not validate '${file.name}`, () => {
+          expect(validate(file.data)).toEqual(false);
+        });
+      });
     });
   });
 });