+++ packages/editor/editor-core/src/nodeviews/index.ts
@@ -1,2 +1,3 @@
 export * from './types';
 export { default as ReactNodeView } from './ReactNodeView';
+export * from './context-adapter';