+++ packages/editor/editor-core/src/plugins/table/nodeviews/TableComponent.tsx
@@ -3,7 +3,11 @@ import rafSchedule from 'raf-schd';
 import { updateColumnsOnResize } from 'prosemirror-tables';
 import { Node as PmNode } from 'prosemirror-model';
 import { EditorView } from 'prosemirror-view';
-import { browser, calcTableWidth } from '@atlaskit/editor-common';
+import {
+  browser,
+  calcTableWidth,
+  akEditorMobileBreakoutPoint,
+} from '@atlaskit/editor-common';
 
 import TableFloatingControls from '../ui/TableFloatingControls';
 import ColumnControls from '../ui/TableFloatingControls/ColumnControls';
@@ -13,7 +17,7 @@ import { scaleTable, setColumnWidths } from '../pm-plugins/table-resizing';
 
 import { TablePluginState, TableCssClassName as ClassName } from '../types';
 import { getCellMinWidth } from '../';
-
+import * as classnames from 'classnames';
 const isIE11 = browser.ie_version === 11;
 
 import { Props } from './table';
@@ -34,6 +38,7 @@ export interface ComponentProps extends Props {
 
   containerWidth: WidthPluginState;
   pluginState: TablePluginState;
+  width: number;
 }
 
 class TableComponent extends React.Component<ComponentProps> {
@@ -122,7 +127,7 @@ class TableComponent extends React.Component<ComponentProps> {
   }
 
   render() {
-    const { view, node, pluginState, containerWidth } = this.props;
+    const { view, node, pluginState, containerWidth, width } = this.props;
     const {
       pluginConfig: { allowControls = true },
     } = pluginState;
@@ -190,15 +195,16 @@ class TableComponent extends React.Component<ComponentProps> {
         style={{
           width: this.getTableContainerWidth(node.attrs.layout, containerWidth),
         }}
-        className={`${ClassName.TABLE_CONTAINER} ${
-          tableActive ? ClassName.WITH_CONTROLS : ''
-        }`}
+        className={classnames(ClassName.TABLE_CONTAINER, {
+          [ClassName.WITH_CONTROLS]: tableActive,
+          'less-padding': width < akEditorMobileBreakoutPoint,
+        })}
         data-number-column={node.attrs.isNumberColumnEnabled}
         data-layout={node.attrs.layout}
       >
         {allowControls && rowControls}
         <div
-          className={ClassName.TABLE_NODE_WRAPPER}
+          className={classnames(ClassName.TABLE_NODE_WRAPPER)}
           ref={elem => {
             this.wrapper = elem;
             this.props.contentDOM(elem ? elem : undefined);