+++ packages/people-and-teams/profilecard/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/profilecard
 
+## 5.0.0
+- [major] [a6dd6e3](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a6dd6e3):
+
+  - Removed isCensored prop as this state no longer exists
+
 ## 4.0.10
 - [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
   - @atlaskit/docs@5.0.8